requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
<<<<<<< HEAD
certifi==2021.5.30
cloudinary==1.26.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap4==3.0.1
django-braces==1.14.0
django-decouple==2.1
django-heroku==0.3.1
django-model-utils==4.1.1
django-registration==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.6
=======
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-heroku==0.3.1
django-registration==3.2
django-tinymce==2.7.0
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
>>>>>>> 770fe889572449248a083b28e7d97e597d32054e
whitenoise==5.3.0
